Refunds implementation
Due to Ethereum EVM limitations, the smart-contract has access only to the last 256 blocks (prof: https://solidity.readthedocs.io/en/v0.5.8/units-and-global-variables.html#block-and-transaction-properties). If for some reason, the transaction with the call of settleBets(...) method is mined 256 blocks after the transaction with the bet, smart-contract has no access to its hash and couldn't determine the winner. In this case we now do the refund.