Batch draws and Superprize implementation
Ethereum EVM has a limit of gas for the methods call included into one block. To avoid this limit overflow on big volumes and make the gas usage controllable we implement the batch draw in settleBets() method. It should increase the smoothness and stability of the game. This commit also includes the Superprize implementation and some rework to reduce the draw time.